Add test for testing similar device
