+++ datadog_checks_base/tests/stubs/test_aggregator_similar.py
@@ -1,8 +1,11 @@
 # (C) Datadog, Inc. 2018
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+import difflib
+
 from datadog_checks.base import AgentCheck
 from datadog_checks.base.stubs import similar
+from datadog_checks.base.stubs.aggregator import AggregatorStub
 from datadog_checks.base.stubs.common import HistogramBucketStub, MetricStub, ServiceCheckStub
 
 
@@ -15,20 +18,21 @@ class TestSimilarAssertionMessages(object):
         check.gauge('test.most_similar_metric', 0)
         check.gauge('test.very_very_different', 0)
 
-        expected_metric = MetricStub("test.similar_metric", None, None, None, None)
-        actual_msg = similar.build_similar_elements_msg(expected_metric, aggregator._metrics)
+        expected_metric = MetricStub("test.similar_metric", None, None, None, None, None)
+        actual_msg = similar.build_similar_elements_msg(expected_metric, aggregator._metrics).strip()
 
         expected_msg = '''
-Expected:
-        MetricStub(name='test.similar_metric', type=None, value=None, tags=None, hostname=None)
+    Expected:
+        MetricStub(name='test.similar_metric', type=None, value=None, tags=None, hostname=None, device=None)
 Similar submitted:
 Score   Most similar
-0.88    MetricStub(name='test.most_similar_metric', type=0, value=0.0, tags=[], hostname='')
-0.83    MetricStub(name='test.another_similar_metric', type=0, value=0.0, tags=[], hostname='')
-0.62    MetricStub(name='test.very_different_metric', type=0, value=0.0, tags=[], hostname='')
-0.42    MetricStub(name='test.very_very_different', type=0, value=0.0, tags=[], hostname='')
-        '''
-        assert expected_msg.strip() == actual_msg.strip(), "Actual message:\n" + actual_msg
+0.88    MetricStub(name='test.most_similar_metric', type=0, value=0.0, tags=[], hostname='', device=None)
+0.83    MetricStub(name='test.another_similar_metric', type=0, value=0.0, tags=[], hostname='', device=None)
+0.62    MetricStub(name='test.very_different_metric', type=0, value=0.0, tags=[], hostname='', device=None)
+0.42    MetricStub(name='test.very_very_different', type=0, value=0.0, tags=[], hostname='', device=None)
+    '''.strip()
+        delta = difflib.ndiff([expected_msg], [actual_msg])
+        assert expected_msg == actual_msg, delta
 
     def test__build_similar_elements__metric_name(self, aggregator):
         check = AgentCheck()
@@ -38,7 +42,9 @@ Score   Most similar
         check.gauge('test.most_similar_metric', 0)
         check.gauge('test.very_very_different', 0)
 
-        expected_metric = MetricStub("test.similar_metric", type=None, value=None, tags=None, hostname=None)
+        expected_metric = MetricStub(
+            "test.similar_metric", type=None, value=None, tags=None, hostname=None, device=None
+        )
         similar_metrics = similar._build_similar_elements(expected_metric, aggregator._metrics)
 
         expected_most_similar_metric = similar_metrics[0][1]
@@ -54,7 +60,7 @@ Score   Most similar
         check.gauge('test.similar_metric2', 20)
         check.gauge('test.similar_metric3', 30)
 
-        expected_metric = MetricStub("test.my_metric", type=None, value=20, tags=None, hostname=None)
+        expected_metric = MetricStub("test.my_metric", type=None, value=20, tags=None, hostname=None, device=None)
         similar_metrics = similar._build_similar_elements(expected_metric, aggregator._metrics)
 
         expected_most_similar_metric = similar_metrics[0][1]
@@ -70,7 +76,7 @@ Score   Most similar
         check.gauge('test.similar_metric3', 10, tags=['something:different'])
 
         expected_metric = MetricStub(
-            "test.test.similar_metric", type=None, value=10, tags=['name:similar_tag'], hostname=None
+            "test.test.similar_metric", type=None, value=10, tags=['name:similar_tag'], hostname=None, device=None
         )
         similar_metrics = similar._build_similar_elements(expected_metric, aggregator._metrics)
 
@@ -87,9 +93,32 @@ Score   Most similar
         check.gauge('test.similar_metric3', 10, hostname='different')
 
         expected_metric = MetricStub(
-            "test.test.similar_metric", type=None, value=10, tags=None, hostname='similar_host'
+            "test.test.similar_metric", type=None, value=10, tags=None, hostname='similar_host', device=None
         )
         similar_metrics = similar._build_similar_elements(expected_metric, aggregator._metrics)
+
+        # expect similar metrics in a similarity order
+        assert similar_metrics[0][1].name == 'test.similar_metric1'
+        assert similar_metrics[1][1].name == 'test.similar_metric2'
+        assert similar_metrics[2][1].name == 'test.similar_metric3'
+
+    def test__build_similar_elements__metric_device(self, aggregator):
+        metrics = {
+            'test.similar_metric2': [
+                MetricStub('test.similar_metric2', AggregatorStub.GAUGE, 10, [], None, 'less_similar_device')
+            ],
+            'test.similar_metric1': [
+                MetricStub('test.similar_metric1', AggregatorStub.GAUGE, 10, [], None, 'similar_device')
+            ],
+            'test.similar_metric3': [
+                MetricStub('test.similar_metric3', AggregatorStub.GAUGE, 10, [], None, 'different')
+            ],
+        }
+
+        expected_metric = MetricStub(
+            "test.test.similar_metric", type=None, value=10, tags=None, hostname=None, device='similar_device'
+        )
+        similar_metrics = similar._build_similar_elements(expected_metric, metrics)
 
         # expect similar metrics in a similarity order
         assert similar_metrics[0][1].name == 'test.similar_metric1'
